Let WordPress handle setting the Cache-Control headers for KeyCDN requests.
KeyCDN supports the `s-max-age` HTTP header which instructs KeyCDN how long to hold on to the response before it is considered stale. Defaults to one day.

Since KeyCDN honors the `s-max-age` header we can instruct browsers to not locally cache the response and check in with the KeyCDN edge node everytime to ensure they are getting the freshest version of the page.

See https://www.keycdn.com/support/frequently-asked-questions/#s-maxage-supported